refactor: reorganize swap-related imports and introduce new swap configurations
- Updated import paths for swap-related components to improve code organization and maintainability.
- Introduced new configuration files for swap settings, including swapConfig and nativeSwapAffiliateConfig, to centralize swap-related parameters.
- Added error handling for swap routes and enhanced the structure of swap quote retrieval functions to support multiple providers.
- Implemented new utility functions for managing native swap assets and quotes, streamlining the swap process across different chains.
