+++ clients/desktop/src/vault/components/VaultPrimaryActions/index.tsx
@@ -1,10 +1,10 @@
 import { CoinKey } from '@core/chain/coin/Coin'
+import { swapEnabledChains } from '@core/chain/swap/swapEnabledChains'
 import { UniformColumnGrid } from '@lib/ui/css/uniformColumnGrid'
 import { ValueProp } from '@lib/ui/props'
 import { isEmpty } from '@lib/utils/array/isEmpty'
 import { isOneOf } from '@lib/utils/array/isOneOf'
 
-import { swapEnabledChains } from '../../../chain/swap/swapEnabledChains'
 import { SendPrompt } from '../../send/SendPrompt'
 import { useCurrentVaultNativeCoins } from '../../state/currentVaultCoins'
 import { SwapPrompt } from '../../swap/components/SwapPrompt'
