+++ clients/desktop/src/vault/keysign/shared/KeysignTxOverview.tsx
@@ -1,5 +1,6 @@
 import { fromChainAmount } from '@core/chain/amount/fromChainAmount'
 import { Chain } from '@core/chain/Chain'
+import { nativeSwapChains } from '@core/chain/swap/native/NativeSwapChain'
 import { formatFee } from '@core/chain/tx/fee/format/formatFee'
 import { getBlockExplorerUrl } from '@core/chain/utils/getBlockExplorerUrl'
 import { fromCommCoin } from '@core/mpc/types/utils/commCoin'
@@ -23,7 +24,6 @@ import { matchDiscriminatedUnion } from '@lib/utils/matchDiscriminatedUnion'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { nativeSwapChains } from '../../../chain/swap/native/NativeSwapChain'
 import { useCopyTxHash } from '../../../chain/ui/hooks/useCopyTxHash'
 import { useCoinPriceQuery } from '../../../coin/query/useCoinPriceQuery'
 import { useFiatCurrency } from '../../../preferences/state/fiatCurrency'
