+++ clients/desktop/src/vault/swap/form/SwapCoinInput.tsx
@@ -1,6 +1,7 @@
 import { CoinKey } from '@core/chain/coin/Coin'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
 import { isNativeCoin } from '@core/chain/coin/utils/isNativeCoin'
+import { swapEnabledChains } from '@core/chain/swap/swapEnabledChains'
 import { Opener } from '@lib/ui/base/Opener'
 import { ChevronDownIcon } from '@lib/ui/icons/ChevronDownIcon'
 import { HStack } from '@lib/ui/layout/Stack'
@@ -12,7 +13,6 @@ import { FC } from 'react'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { swapEnabledChains } from '../../../chain/swap/swapEnabledChains'
 import { ChainCoinIcon } from '../../../chain/ui/ChainCoinIcon'
 import { getChainEntityIconSrc } from '../../../chain/utils/getChainEntityIconSrc'
 import { getCoinLogoSrc } from '../../../coin/logo/getCoinLogoSrc'
