+++ clients/desktop/src/vault/swap/form/info/SwapProvider.tsx
@@ -1,8 +1,8 @@
+import { getSwapQuoteProviderName } from '@core/chain/swap/quote/getSwapQuoteProviderName'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { Text } from '@lib/ui/text'
 import { useTranslation } from 'react-i18next'
 
-import { getSwapQuoteProviderName } from '../../../../chain/swap/quote/getSwapQuoteProviderName'
 import { Skeleton } from '../../../../components/skeleton'
 import { StrictInfoRow } from '../../../../lib/ui/layout/StrictInfoRow'
 import { useSwapQuoteQuery } from '../../queries/useSwapQuoteQuery'
