+++ clients/desktop/src/vault/swap/form/info/SwapTotalFeeFiatValue.tsx
@@ -1,9 +1,9 @@
+import { SwapFee } from '@core/chain/swap/SwapFee'
 import { ValueProp } from '@lib/ui/props'
 import { MatchEagerQuery } from '@lib/ui/query/components/MatchEagerQuery'
 
 import { Skeleton } from '../../../../components/skeleton'
 import { useSwapFiatFeesQuery } from '../../queries/useSwapFiatFeesQuery'
-import { SwapFee } from '../../types/SwapFee'
 
 export const SwapFeeFiatValue = ({ value }: ValueProp<SwapFee[]>) => {
   const pricesQuery = useSwapFiatFeesQuery(value)
