+++ clients/desktop/src/vault/swap/queries/useSwapChainSpecificQuery.ts
@@ -2,13 +2,13 @@ import { toChainAmount } from '@core/chain/amount/toChainAmount'
 import { UtxoChain } from '@core/chain/Chain'
 import { chainFeeCoin } from '@core/chain/coin/chainFeeCoin'
 import { areEqualCoins } from '@core/chain/coin/Coin'
+import { getSwapKeysignPayloadFields } from '@core/chain/swap/keysign/getSwapKeysignPayloadFields'
 import { getChainSpecific } from '@core/mpc/keysign/chainSpecific'
 import { ChainSpecificResolverInput } from '@core/mpc/keysign/chainSpecific/ChainSpecificResolver'
 import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 import { useStateDependentQuery } from '@lib/ui/query/hooks/useStateDependentQuery'
 import { isOneOf } from '@lib/utils/array/isOneOf'
 
-import { getSwapKeysignPayloadFields } from '../../../chain/swap/keysign/getSwapKeysignPayloadFields'
 import { toHexPublicKey } from '../../../chain/utils/toHexPublicKey'
 import { getChainSpecificQueryKey } from '../../../coin/query/useChainSpecificQuery'
 import { useVaultPublicKeyQuery } from '../../publicKey/queries/useVaultPublicKeyQuery'
