+++ clients/desktop/src/vault/swap/queries/useSwapFeesQuery.ts
@@ -1,4 +1,5 @@
 import { chainFeeCoin } from '@core/chain/coin/chainFeeCoin'
+import { SwapFees } from '@core/chain/swap/SwapFee'
 import { getFeeAmount } from '@core/chain/tx/fee/getFeeAmount'
 import { useTransformQueriesData } from '@lib/ui/query/hooks/useTransformQueriesData'
 import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
@@ -6,7 +7,6 @@ import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
 import { useCurrentVaultCoin } from '../../state/currentVaultCoins'
 import { useFromCoin } from '../state/fromCoin'
 import { useToCoin } from '../state/toCoin'
-import { SwapFees } from '../types/SwapFee'
 import { useSwapChainSpecificQuery } from './useSwapChainSpecificQuery'
 import { useSwapQuoteQuery } from './useSwapQuoteQuery'
