+++ clients/desktop/src/vault/swap/queries/useSwapOutputAmountQuery.ts
@@ -1,11 +1,11 @@
 import { fromChainAmount } from '@core/chain/amount/fromChainAmount'
+import { GeneralSwapQuote } from '@core/chain/swap/general/GeneralSwapQuote'
+import { getNativeSwapDecimals } from '@core/chain/swap/native/utils/getNativeSwapDecimals'
+import { SwapQuote } from '@core/chain/swap/quote/SwapQuote'
 import { useTransformQueryData } from '@lib/ui/query/hooks/useTransformQueryData'
 import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
 import { useCallback } from 'react'
 
-import { GeneralSwapQuote } from '../../../chain/swap/general/GeneralSwapQuote'
-import { getNativeSwapDecimals } from '../../../chain/swap/native/utils/getNativeSwapDecimals'
-import { SwapQuote } from '../../../chain/swap/quote/SwapQuote'
 import { useCurrentVaultCoin } from '../../state/currentVaultCoins'
 import { useToCoin } from '../state/toCoin'
 import { useSwapQuoteQuery } from './useSwapQuoteQuery'
