+++ core/chain/swap/general/oneInch/api/getOneInchSwapQuote.ts
@@ -2,6 +2,8 @@ import { EvmChain } from '@core/chain/Chain'
 import { ChainAccount } from '@core/chain/ChainAccount'
 import { getEvmChainId } from '@core/chain/chains/evm/chainInfo'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
+import { OneInchSwapQuoteResponse } from '@core/chain/swap/general/oneInch/api/OneInchSwapQuoteResponse'
+import { oneInchAffiliateConfig } from '@core/chain/swap/general/oneInch/oneInchAffiliateConfig'
 import { defaultEvmSwapGasLimit } from '@core/chain/tx/fee/evm/evmGasLimit'
 import { rootApiUrl } from '@core/config'
 import { addQueryParams } from '@lib/utils/query/addQueryParams'
@@ -9,8 +11,6 @@ import { queryUrl } from '@lib/utils/query/queryUrl'
 import { pick } from '@lib/utils/record/pick'
 
 import { GeneralSwapQuote } from '../../GeneralSwapQuote'
-import { oneInchAffiliateConfig } from '../oneInchAffiliateConfig'
-import { OneInchSwapQuoteResponse } from './OneInchSwapQuoteResponse'
 
 type Input = {
   account: ChainAccount
