+++ core/chain/swap/native/thor/tx/getThorchainSwapTxInputData.ts
@@ -1,5 +1,6 @@
 import { Chain } from '@core/chain/Chain'
 import { getSigningInputEnvelopedTxFields } from '@core/chain/chains/evm/tx/getSigningInputEnvelopedTxFields'
+import { toKeysignSwapPayload } from '@core/mpc/keysign/swap/KeysignSwapPayload'
 import { fromCommCoin } from '@core/mpc/types/utils/commCoin'
 import { EthereumSpecific } from '@core/mpc/types/vultisig/keysign/v1/blockchain_specific_pb'
 import { KeysignPayload } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
@@ -9,7 +10,6 @@ import { match } from '@lib/utils/match'
 import { TW, WalletCore } from '@trustwallet/wallet-core'
 import Long from 'long'
 
-import { toKeysignSwapPayload } from '../../../../keysign/KeysignSwapPayload'
 import { nativeSwapAffiliateConfig } from '../../nativeSwapAffiliateConfig'
 import { toThorchainSwapAssetProto } from '../asset/toThorchainSwapAssetProto'
 import { ThorchainSwapEnabledChain } from '../thorchainSwapProtoChains'
