+++ core/chain/swap/native/thor/utils/thorchainSwapQuoteToSwapPayload.ts
@@ -1,12 +1,12 @@
 import { create } from '@bufbuild/protobuf'
 import { fromChainAmount } from '@core/chain/amount/fromChainAmount'
 import { AccountCoin } from '@core/chain/coin/AccountCoin'
+import { KeysignSwapPayload } from '@core/mpc/keysign/swap/KeysignSwapPayload'
 import { toCommCoin } from '@core/mpc/types/utils/commCoin'
 import { THORChainSwapPayloadSchema } from '@core/mpc/types/vultisig/keysign/v1/thorchain_swap_payload_pb'
 import { convertDuration } from '@lib/utils/time/convertDuration'
 import { addMinutes } from 'date-fns'
 
-import { KeysignSwapPayload } from '../../../../keysign/KeysignSwapPayload'
 import { nativeSwapStreamingInterval } from '../../NativeSwapChain'
 import { NativeSwapQuote } from '../../NativeSwapQuote'
