+++ core/chain/swap/native/utils/getNativeSwapDecimals.ts
