+++ core/chain/swap/swapEnabledChains.ts
@@ -1,7 +1,6 @@
 import { lifiSwapEnabledChains } from '@core/chain/swap/general/lifi/LifiSwapEnabledChains'
 import { oneInchSwapEnabledChains } from '@core/chain/swap/general/oneInch/OneInchSwapEnabledChains'
-
-import { nativeSwapEnabledChains } from './native/NativeSwapChain'
+import { nativeSwapEnabledChains } from '@core/chain/swap/native/NativeSwapChain'
 
 export const swapEnabledChains = [
   ...nativeSwapEnabledChains,
